third_party/WebKit/LayoutTests/fast/inspector-support/uncaught-dom3-exception-expected.txt

CONSOLE MESSAGE: line 9: Uncaught HierarchyRequestError: Failed to execute 'appendChild' on 'Node': Nodes of type 'UNDEFINED' may not be inserted inside nodes of type '#document'.
This tests that we are getting the correct message for DOM Exception 3: HIERARCHY_REQUEST_ERR.
